Improve Dockerfile
This commit "pin points" several aspects of building the Docker image
for ping-exporter to explicit versions:

* Use alpine-3.8, both for building as well as for running ping-exporter
* Do NOT use a "install.sh" script, fetched from the internets and piped
  to sh to get "dep" (one of the Golang dependency fetchers).
* Instead, use the version that is available as an Alpine package. As an
  additional benefit: "dep" does not have to be compiled and thus
  building ping-exporter is a bit faster.
* Use Golang 1.11 as the effective Golang compiler to build
  ping-exporter
* Increase verbosity while building the image

Last, this commit changes the behavior of the Docker image to NOT provide an
"empty" configuration file for ping-exporter. The reason for this is to
let ping-exporter fail with a meaningful error message. When an empty
configuration file is read (and ping-exporter reads that file just
fine), the next error message consists of complaints about not providing
web.listen-address etc. What the users actually want to do is to volume
mount a working configuration file into a container. If they forgot
about doing that: ping-exporter will complaint about a missing file. If
they don't want to use that file alltogether: --config "" is there for
rescue.